Made C++ version compile on Windows.

+++ source/quake/view.cpp
@@ -278,7 +278,7 @@ void BuildGammaTable (float g)
 	
 	for (i=0 ; i<256 ; i++)
 	{
-		inf = 255 * pow ( (i+0.5)/255.5 , g ) + 0.5;
+		inf = 255 * pow( (i+0.5)/255.5 , (double)g ) + 0.5;
 		if (inf < 0)
 			inf = 0;
 		if (inf > 255)
